storages_compare/requirements.txt
annotated-types==0.7.0
cffi==1.16.0
cryptography==42.0.7
dnspython==2.6.1
Faker==25.3.0
mypy==1.10.0
mypy-extensions==1.0.0
psycopg2==2.9.9
pycparser==2.22
pydantic==2.7.2
pydantic-settings==2.3.0
pydantic_core==2.18.3
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
ruff==0.4.7
six==1.16.0
types-cffi==1.16.0.20240331
types-psycopg2==2.9.21.20240417
types-pyOpenSSL==24.1.0.20240425
types-PyYAML==6.0.12.20240311
types-requests==2.32.0.20240602
types-setuptools==70.0.0.20240524
typing_extensions==4.12.1
urllib3==2.2.1
